Container/Webinterface/requirements.txt
rpi.gpio
Flask
flask_socketio
